requirements.txt
mock==1.0.1
py==1.4.27
pyGravatar==0.0.6
pytest==2.7.1
requests==2.7.0
